UI and slider value when click
