--- Aim For That/Aim For That/ViewController.swift
@@ -1,28 +1,0 @@
-//
-//  ViewController.swift
-//  Aim For That
-//
-//  Created by Miguel Herrero on 21/7/17.
-//  Copyright © 2017 Miguel Herrero. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-    @IBAction func showAlert() {
-        print("He pulsado el botón")
-    }
-    
-}
-
